update
new screenshots and added some more clear text what is what in the ble communication. Noticed inconsistency how exactly the communication is being made between the rpi and esp32. Added more clearer text and changed the name, since in my head the rpi acts as the so called server since it forwards the data but it does so as a client in the ble side of things.